Several updates... - Change the default small model - Set up pipelining - Ensure that PIL images are saved to buffers and returned correctly - Add generation options that ( hopefully ) align with machine   good-itude, and pipe them through the generation functions   These are operated based on environment vars so far.
It generated an image -- Yay\! We need to add more models to the list.
It might be beneficial to add them as defaults to the default options
for the different machine types. We'll see.

Also, We should bottom out at 512x512 for SD. It was apparently trained
on that size, and below that gives you eldritch abominations.